Fixed all implicit type casting errors for CI build fail.

+++ examples/kinsol/CXX_parallel/kin_heat2D_nonlin_p.hpp
@@ -130,7 +130,7 @@ struct UserData
   // Fixed Point Solver settings
   realtype rtol;        // relative tolerance
   int      maa;         // m for Anderson Acceleration
-  int      damping;     // daming for Anderson Acceleration
+  double   damping;     // daming for Anderson Acceleration
   int      orthaa;      // orthogonalization routine for AA
   int      maxits;      // max number of fixed point iterations
